+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Add autosave feature for `blueberry4$` users
 * Add `unpause_eggs` feature
+* Turn "Toggle Powerups" into "Enable Powerups"
 
 ## 0.1.3 (2018-03-17)